+++ app/Http/Controllers/UserListsItemController.php
@@ -5,11 +5,14 @@ namespace App\Http\Controllers;
 use App\Http\Requests\UserListsItem\UserListsItemCreateRequest;
 use App\Http\Requests\UserListsItem\UserListsItemDeleteRequest;
 use App\Http\Requests\UserListsItem\UserListsItemUpdateRequest;
+use App\Modules\UserList\Domain\Entities\UserListEntity;
 use App\Modules\UserListItem\Application\Manager\UserListItemManager;
 use App\Modules\UserListItem\Domain\DTO\UserListItemDTO;
+use App\Modules\UserListItem\Domain\Entities\UserListsItemEntity;
 use Exception;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 
 class UserListsItemController extends Controller
 {
@@ -58,6 +61,8 @@ class UserListsItemController extends Controller
      */
     public function update(UserListsItemUpdateRequest $request): JsonResponse
     {
+        Gate::authorize('isOwner', [new UserListsItemEntity(), $request->list_item_id]);
+
         try {
             $userListDTO = UserListItemDTO::fromUpdateRequest($request);
             $userListItem = $this->userListItemManager->update($request->list_item_id, $userListDTO);
@@ -80,6 +85,8 @@ class UserListsItemController extends Controller
      */
     public function delete(UserListsItemDeleteRequest $request): JsonResponse
     {
+        Gate::authorize('isOwner', [new UserListsItemEntity(), $request->list_item_id]);
+
         try {
             $this->userListItemManager->delete($request->list_item_id);
             return response()->json(["message" => "User list item deletion is successfuly completed."], 200);
